updated release.txt, rxr optimization (no file buffering, write to every page at memory allocation, fix to fifo depth

+++ slsReceiverSoftware/src/UDPStandardImplementation.cpp
@@ -836,7 +836,9 @@ int UDPStandardImplementation::SetupFifoStructure() {
 		if(dataStreamer.size())dataStreamer[i]->SetFifo(fifo[i]);
 	}
 
-	FILE_LOG(logINFO) << "Memory Allocated Per Fifo: " << (double)( ((generalData->imageSize) * numberofJobs + (generalData->fifoBufferHeaderSize)) * fifoDepth)/ (double)(1024 * 1024) << " MB" ;
+	FILE_LOG(logINFO) << "Memory Allocated Per Fifo: " << (double)( ((size_t)(generalData->imageSize) * numberofJobs + 
+	(size_t)(generalData->fifoBufferHeaderSize)) * (size_t)fifoDepth)/ 
+	(double)(1024 * 1024) << " MB" ;
 	FILE_LOG(logINFO) << numThreads << " Fifo structure(s) reconstructed";
 	return OK;
 }
